+++ bin/scripts/app.js
@@ -100,16 +100,18 @@ define('app/events/EventConstants',
 define('app/events/EventCreator',
   function (require, module, exports) {
 
-    var _dispatcher     = require('nori/utils/Dispatcher'),
-        _eventConstants = require('app/events/EventConstants');
+    var _eventConstants = require('app/events/EventConstants');
 
     module.exports.someEvent = function (data) {
-      _dispatcher.publish({
+      var evtObj = {
         type   : _eventConstants.SOMETHING_HAPPENED,
         payload: {
           theData: data
         }
-      });
+      };
+
+      Nori.dispatcher().publish(evtObj);
+      return evtObj;
     };
 
   });