+++ src/scripts/nori/Nori.js
@@ -370,6 +370,7 @@ var Nori = (function () {
     router                : getRouter,
     model                 : getModel,
     view                  : getView,
+    dispatcher            : getDispatcher,
     createApplication     : createApplication,
     createApplicationModel: createApplicationModel,
     createApplicationView : createApplicationView,